feat(visualizer): collapse toolbar into single top row (req #2407)
